[MC-36] Created `utils.py` in tests.
